+++ configs/base_config.yaml
@@ -1,13 +1,9 @@
 # 데이터 및 훈련 관련 설정
 data :
-  train_data_path : '/data/ephemeral/home/data/train/DCM'
-  train_label_path : '/data/ephemeral/home/data/train/outputs_json'
-  test_data_path : '/data/ephemeral/home/data/test/DCM'
-  # classes : [
-  #   'Trapezium',
-  #   'Trapezoid', 'Capitate', 'Hamate', 'Scaphoid', 'Lunate',
-  #   'Triquetrum', 'Pisiform', 'Radius', 'Ulna'
-  # ]
+  train_data_path : '../../data/train/DCM'
+  train_label_path : '../../data/train/outputs_json'
+  test_data_path : '../../data/test/DCM'
+
 
   classes : [
     'finger-1', 'finger-2', 'finger-3', 'finger-4', 'finger-5',
@@ -20,6 +16,7 @@ data :
 
   fold : 0
   split_method : GroupKFold
+  splits : 5
 
 
   train : 
@@ -42,17 +39,9 @@ data :
     num_workers : 8
     pin_memory : True
     drop_last : False
-    interval: 10  # validation 및 체크포인트 저장 주기
+    interval: 10  
     threshold: 0.5
     check: True
-
-    interpolate:
-      nearest: 'nearest'
-      bilinear: 'bilinear'
-      bicubic: 'bicubic'
-      area: 'area'
-      lanczos: 'lanczos'
-
   
 
   test : 
@@ -61,6 +50,24 @@ data :
     num_workers : 8
     pin_memory : True
     drop_last : False
+
+
+
+inference:
+  threshold: 0.5
+  output_size: [2048, 2048] 
+  checkpoint_path: "/data/ephemeral/home/level2-cv-semanticsegmentation-cv-22-lv3/checkpoints/lraspp_mobilenet/lraspp_mobilenet_v3_large_best_model.pt"  # .ckpt에서 .pt로 수정
+  output_dir: "/data/ephemeral/home/level2-cv-semanticsegmentation-cv-22-lv3/results"
+  mode: 'gpu'
+
+
+
+interpolate:
+    nearest: 'nearest'
+    bilinear: 'bilinear'
+    bicubic: 'bicubic'
+    area: 'area'
+    lanczos: 'lanczos'
   
 
 augmentation:
@@ -99,22 +106,10 @@ augmentation:
         mean: [0.485, 0.456, 0.406]
         std: [0.229, 0.224, 0.225]
 
-save:
-  # 실험 config 저장 경로
-  save_config: /data/ephemeral/home/level2-cv-semanticsegmentation-cv-22-lv3/configs
-  # 체크포인트 저장 경로
-  save_ckpt: /data/ephemeral/home/level2-cv-semanticsegmentation-cv-22-lv3/checkpoints
-
-
-seed: 42
-
 
 loss_func: 
-  weight_map : True
-  # BCEWithLogitsLoss, DiceLoss, JaccardLoss, FocalLoss,  
-  # CombinedWeightedLoss, TwoWayLoss, BCEDiceLoss
-  name: DiceLoss
-
+  weight_map : None
+  name: CombinedWeightedLoss # BCEWithLogitsLoss, DiceLoss, JaccardLoss, FocalLoss,  CombinedWeightedLoss, TwoWayLoss, BCEDiceLoss
 
 loss_func_defaults:
   BCEWithLogitsLoss: {}
@@ -136,9 +131,9 @@ loss_func_defaults:
     dice_mode: multilabel
 
 
+
 optimizer: 
-# Adam, SGD, AdamW, RMSprop, Adagrad, Adadelta
-  name: AdamW
+  name: AdamW # Adam, SGD, AdamW, RMSprop, Adagrad, Adadelta
   params:
     lr: 1e-4
     weight_decay: 1e-2
@@ -156,34 +151,33 @@ optimizer_defaults:
     weight_decay: 1e-6
   AdamW:
     lr: 1e-4
-    weight_decay: 1e-2  # AdamW는 weight_decay를 상대적으로 더 크게 설정하는 것이 일반적
+    weight_decay: 1e-2  
   RMSprop:
     lr: 1e-4
-    alpha: 0.99        # smoothing constant
+    alpha: 0.99       
     momentum: 0.9
     weight_decay: 1e-6
   Adagrad:
-    lr: 1e-2           # Adagrad는 보통 비교적 큰 학습률로 시작
+    lr: 1e-2          
     weight_decay: 1e-6
     initial_accumulator_value: 0.1
   Adadelta:
-    lr: 1.0            # Adadelta는 보통 기본값이 1.0
-    rho: 0.9           # decay rate
+    lr: 1.0            
+    rho: 0.9          
     eps: 1e-6
-    weight_decay: 0    # 일반적으로 Adadelta는 weight_decay를 사용하지 않음
+    weight_decay: 0   
 
 
 scheduler:
-  # CosineAnnealingLR, StepLR, ReduceLROnPlateau, ExponentialLR, MultiStepLR, CosineAnnealingWarmRestarts
-  name: CosineAnnealingLR
+  name: CosineAnnealingLR   # CosineAnnealingLR, StepLR, ReduceLROnPlateau, ExponentialLR, MultiStepLR, CosineAnnealingWarmRestarts
   params:
     T_max: 50
 
 scheduler_defaults:
   CosineAnnealingLR:
     T_max: 100
-    eta_min: 0        # default
-    last_epoch: -1    # default
+    eta_min: 0       
+    last_epoch: -1   
   StepLR:
     step_size: 20
     gamma: 0.1
@@ -200,26 +194,20 @@ scheduler_defaults:
     gamma: 0.1
     last_epoch: -1
   MultiStepLR:
-    # 특정 에폭에서 학습률을 감소
     milestones: [30, 60, 90]
     gamma: 0.1
     last_epoch: -1
   CosineAnnealingWarmRestarts:
-    # 학습률을 코사인 곡선을 따라 감소시키다가 설정된 주기마다 다시 초기 학습률로 "재시작"하는 방식
-    T_0: 10           # 첫 번째 주기의 길이
-    T_mult: 1         # 각 주기마다 늘리는 비율
-    eta_min: 0        # 학습률이 도달할 수 있는 최저값
-    last_epoch: -1    # 스케줄러가 몇 번째 epoch에서 시작해야 하는지 
+    T_0: 10          
+    T_mult: 1        
+    eta_min: 0       
+    last_epoch: -1    
 
 
-# 디버깅 모드
+
 debug: True
+save:
+  save_config: /data/ephemeral/home/level2-cv-semanticsegmentation-cv-22-lv3/configs
+  save_ckpt: /data/ephemeral/home/level2-cv-semanticsegmentation-cv-22-lv3/checkpoints
 
-
-# Inference 관련 설정
-inference:
-  threshold: 0.5
-  output_size: [2048, 2048]  # 튜플을 리스트로 수정
-  checkpoint_path: "/data/ephemeral/home/level2-cv-semanticsegmentation-cv-22-lv3/checkpoints/lraspp_mobilenet/lraspp_mobilenet_v3_large_best_model.pt"  # .ckpt에서 .pt로 수정
-  output_dir: "/data/ephemeral/home/level2-cv-semanticsegmentation-cv-22-lv3/results"
-  mode: 'gpu'
+seed: 42
