+++ src/inference.py
@@ -1,71 +1,15 @@
-import os
-import pytz
 import torch
 import argparse
-import numpy as np
-import pandas as pd
 from tqdm import tqdm
-from datetime import datetime
 import torch.nn.functional as F
 from omegaconf import OmegaConf
 from train.trainer import set_seed
-import segmentation_models_pytorch as smp
 from model.model_loader import model_loader
 from Dataset.dataloader import get_test_loader
 from model.utils.model_output import get_model_output
+from utils.inference_utils import load_model, encode_mask_to_rle, encode_mask_to_rle_gpu, save_to_csv
 
-def encode_mask_to_rle(mask):
-    pixels = mask.flatten()
-    pixels = np.concatenate([[0], pixels, [0]])
-    runs = np.where(pixels[1:] != pixels[:-1])[0] + 1
-    runs[1::2] -= runs[::2]
-    return ' '.join(str(x) for x in runs)
 
-def encode_mask_to_rle_gpu(mask):
-    mask = mask.to(torch.uint8)
-    pixels = mask.flatten()
-    zeros = torch.zeros(1, dtype=pixels.dtype, device=pixels.device)
-    pixels = torch.cat([zeros, pixels, zeros])
-    changes = (pixels[1:] != pixels[:-1]).nonzero().squeeze() + 1
-    runs = changes.clone()
-    runs[1::2] = runs[1::2] - runs[::2]
-    runs = runs.cpu().numpy()
-    return ' '.join(str(x) for x in runs)
-
-def save_to_csv(filename_and_class, rles, cfg):
-    os.makedirs(cfg.inference.output_dir, exist_ok=True)
-    kst = pytz.timezone('Asia/Seoul')
-    timestamp = datetime.now(kst).strftime("%Y%m%d_%H%M%S")
-    pt_name = cfg.inference.checkpoint_path.split('/')[-1].split('.')[0]
-    output_filepath = os.path.join(cfg.inference.output_dir, f'{pt_name}_{timestamp}.csv')
-
-    classes, filename = zip(*[x.split("_") for x in filename_and_class])
-    image_name = [os.path.basename(f) for f in filename]
-
-    df = pd.DataFrame({
-        "image_name": image_name,
-        "class": classes,
-        "rle": rles,
-    })
-
-    df.to_csv(output_filepath, index=False)
-    print(f"Submission file saved to: {output_filepath}")
-    return output_filepath
-
-def load_model(model, checkpoint_path, device, library):
-    checkpoint = torch.load(checkpoint_path, map_location=device) if library == 'torchvision' else smp.from_pretrained(checkpoint_path)
-
-    if isinstance(checkpoint, dict):
-        if 'state_dict' in checkpoint:
-            model.load_state_dict(checkpoint['state_dict'])
-        else:
-            model.load_state_dict(checkpoint, strict=False)
-    else:
-        model = checkpoint
-
-    print(f"Checkpoint loaded successfully from: {checkpoint_path}")
-    return model
-    
 def do_inference(cfg):
     set_seed(cfg.seed)
     library = cfg.model.library
@@ -122,6 +66,7 @@ def do_inference(cfg):
     except Exception as e:
         print(f"Error during inference: {e}")
         raise
+
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Test Semantic Segmentation Model")
