+++ src/utils/set_seed.py
@@ -2,7 +2,6 @@ import torch
 import random
 import numpy as np
 
-
 def set_seed(seed):
     torch.manual_seed(seed)
     torch.cuda.manual_seed(seed)
